Move DOM lib to lib file.
