+++ runtime/src/runtime.js
@@ -1,10 +1,10 @@
 import { heapAdjust } from './core/heap';
 
 import * as libStrings from './lib/strings';
+import * as libDOM from './lib/DOM';
 
 let grainInitialized = false;
 let grainModule;
-let grainDOMRefs = [];
 
 function debugPrint(n) {
   // console.log(`0x${n.toString(16)} (0b${n.toString(2)})`);
@@ -55,48 +55,6 @@ function grainEqual(x, y) {
   return grainEqualHelp(x, y, 0) ? GRAIN_TRUE : GRAIN_FALSE;
 }
 
-function grainDOMQuery(n) {
-  assertString(n);
-  let query = grainToJSVal(n);
-  let elem = document.querySelector(query);
-  if (elem) {
-    grainDOMRefs.push(elem);
-    let heapRef = grainHeapAllocate(2) / 4;
-    view[heapRef] = GRAIN_DOM_ELEM_TAG;
-    view[heapRef+1] = grainDOMRefs.length - 1;
-    return (heapRef * 4) ^ 3;
-  } else {
-    return GRAIN_FALSE;
-  }
-}
-
-function grainDOMElemSetText(elemRef, textRef) {
-  assertDOMElement(elemRef);
-  assertString(textRef);
-  let elem = grainToJSVal(elemRef);
-  elem.innerText = grainToJSVal(textRef);
-  return elemRef;
-}
-
-function grainDOMDangerouslySetInnerHTML(elemRef, textRef) {
-  assertDOMElement(elemRef);
-  assertString(textRef);
-  let elem = grainToJSVal(elemRef);
-  elem.innerHTML = grainToJSVal(textRef);
-  return elemRef;
-}
-
-function grainDOMAddEventListener(elemRef, eventRef, handlerRef) {
-  assertDOMElement(elemRef);
-  assertString(eventRef);
-  assertLambda(handlerRef);
-  let elem = grainToJSVal(elemRef);
-  let event = grainToJSVal(eventRef);
-  let handler = grainToJSVal(handlerRef);
-  elem.addEventListener(event, () => handler.call());
-  return elemRef;
-}
-
 function displayOnPage(str) {
   document.getElementById('output').innerText = str;
 }
@@ -125,10 +83,7 @@ const importObj = {
     equal: grainEqual,
     toString: (n => JSToGrainVal(grainToString(n))),
     ...libStrings,
-    DOMQuery: grainDOMQuery,
-    DOMSetText: grainDOMElemSetText,
-    DOMDangerouslySetInnerHTML: grainDOMDangerouslySetInnerHTML,
-    DOMAddEventListener: grainDOMAddEventListener
+    ...libDOM
   }
 };
